feat(epub): apply the rules a book hides inside @media, and finish the review (#251)
A great deal of Calibre and KindleGen output wraps its body styling in
`@media screen { … }`. The selector engine skips at-rules wholesale, so such a
book declared nothing at all — margins, alignment, breaks, the lot — and #251's
item 3 would have read as unfixed for it. Flattening beats teaching the engine
media queries: inkread renders to exactly one medium, so a query is not a
condition to evaluate but a question of whether the block is addressed to us. A
query naming print, speech or a vendor medium is not; anything else is. Other
at-rules pass through untouched, and comments are stripped first so a
commented-out block cannot be flattened into live rules.

The module doc claimed `:last-child` and `:only-child` were answered. They are
not, and cannot be: simplecss 0.2.2 has no variant for them, so a rule using one
is dropped before it reaches us. The doc now says that, and names it as the real
gap it is rather than implying coverage the code does not have.

Readability, from the second review: `row_stages` gains a `Flow` alias — the row
machinery nests it two deep in two different orders, cells-of-segments in and
stages-of-cells out, and three anonymous `Vec` levels are a puzzle where two
named ones are not — a signature line saying which order is which, a worked
example, and two extractions (`stage_heights`, `carry_across`) that leave it as
dispatch. `layout_tests.rs` had grown to 2283 lines, twice the guideline and the
largest file in the crate; the #251 property tests move to
`style_layout_tests.rs` with the helpers the two halves had been duplicating
hoisted once.
